# Micro optimize encoding checks